+++ frontend/src/app/analysis-logs/loading.tsx
@@ -1,0 +1,14 @@
+export default function Loading() {
+    return (
+        <div className="min-h-screen bg-gray-50 py-8 px-4">
+            <div className="max-w-4xl mx-auto">
+                <div className="flex items-center justify-center py-12">
+                    <div className="text-center">
+                        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-600 mx-auto mb-4"></div>
+                        <div className="text-gray-600">ログを読み込み中...</div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    );
+}
