+++ frontend/src/app/analysis-logs/page.tsx
@@ -1,5 +1,9 @@
-import AnalysisLogsPage from "./AnalysisLogsPage";
+import { getAnalysisLogs } from '@/lib/api';
+import AnalysisLogsPage from './AnalysisLogsPage';
 
-export default function Page() {
-    return <AnalysisLogsPage />;
+// メインページ（サーバーコンポーネント）
+export default async function Page() {
+    const data = await getAnalysisLogs(1, 20);
+
+    return <AnalysisLogsPage data={data} />;
 }
